# feat(documents): optimise linked/related docs loading