--- a/packages/documents/graphql/resolvers/Document.js
+++ b/packages/documents/graphql/resolvers/Document.js
@@ -17,6 +17,11 @@ const { lib: { webp: {
   addSuffix: addWebpSuffix
 } } } = require('@orbiting/backend-modules-assets')
 
+const {
+  extractIdsFromNode,
+  loadLinkedMetaData
+} = require('@orbiting/backend-modules-search/lib/Documents')
+
 const shouldDeliverWebP = (argument = 'auto', req) => {
   if (argument === 'auto') {
     return req && req.get('Accept').indexOf('image/webp') > -1
@@ -53,7 +58,7 @@ module.exports = {
     }
     return meta
   },
-  children (doc, { first, last, before, after, only, urlPrefix, searchString, webp }, context, info) {
+  async children (doc, { first, last, before, after, only, urlPrefix, searchString, webp }, context, info) {
     if (!doc || !doc.content || !doc.content.children) {
       return {
         pageInfo: {
@@ -67,8 +72,6 @@ module.exports = {
       }
     }
     if (doc._all) {
-      contentUrlResolver(doc, doc._all, doc._usernames, undefined, urlPrefix, searchString, context.user || null)
-
       if (shouldDeliverWebP(webp, context.req)) {
         processRepoImageUrlsInContent(doc.content, addWebpSuffix)
         processImageUrlsInContent(doc.content, addWebpSuffix)
@@ -103,6 +106,21 @@ module.exports = {
       children.some((v, i) => v.data.id === endCursor && i < lastIndex)
     const hasPreviousPage = !!startCursor &&
       children.some((v, i) => v.data.id === startCursor && i > firstIndex)
+
+    if (doc._all) {
+      const idsFromNodes = nodes.map(
+        node => extractIdsFromNode(node, doc.meta.repoId)
+      )
+      const { docs, usernames } = await loadLinkedMetaData({
+        context,
+        userIds: idsFromNodes.reduce((userIds, idsFromNode) => userIds.concat(idsFromNode.users), []),
+        repoIds: idsFromNodes.reduce((repoIds, idsFromNode) => repoIds.concat(idsFromNode.repos), [])
+      })
+      doc._all = doc._all.concat(docs)
+      doc._usernames = doc._usernames.concat(usernames)
+
+      contentUrlResolver(doc, doc._all, doc._usernames, undefined, urlPrefix, searchString, context.user || null)
+    }
 
     return {
       pageInfo: {