--- a/packages/documents/graphql/resolvers/_queries/document.js
+++ b/packages/documents/graphql/resolvers/_queries/document.js
@@ -1,12 +1,12 @@
 const getDocuments = require('./documents')
 
-module.exports = async (_, args, context) => {
+module.exports = async (_, args, context, info) => {
   const {
     id,
     path,
     repoId
   } = args
 
-  return getDocuments(_, { id, path, repoId }, context)
+  return getDocuments(_, { id, path, repoId }, context, info)
     .then(docCon => docCon.nodes[0])
 }