--- a/packages/documents/graphql/resolvers/_queries/documents.js
+++ b/packages/documents/graphql/resolvers/_queries/documents.js
@@ -1,7 +1,7 @@
 const _ = require('lodash')
 const search = require('@orbiting/backend-modules-search/graphql/resolvers/_queries/search')
 
-module.exports = async (__, args, context) => {
+module.exports = async (__, args, context, info) => {
   const docsConnection = await search(null, {
     first: args.first,
     after: args.after,
@@ -14,7 +14,7 @@ module.exports = async (__, args, context) => {
       key: 'publishedAt',
       direction: 'DESC'
     }
-  }, context)
+  }, context, info)
 
   // transform SearchConnection to DocumentConnection
   return {