--- a/packages/search/graphql/resolvers/_queries/search.js
+++ b/packages/search/graphql/resolvers/_queries/search.js
@@ -161,7 +161,7 @@ const createHighlight = (indicesList) => {
 
 const defaultExcludes = [ 'contentString', 'resolved' ]
 const createQuery = (
-  searchTerm, filter, sort, indicesList, user, scheduledAt, withoutContent, withoutAggs, ignorePrepublished
+  searchTerm, filter, sort, indicesList, user, scheduledAt, withoutChildren, withoutAggs, ignorePrepublished
 ) => ({
   query: {
     bool: {
@@ -177,7 +177,7 @@ const createQuery = (
   _source: {
     'excludes': [
       ...defaultExcludes,
-      ...withoutContent
+      ...withoutChildren
         ? [ 'content.children' ]
         : [ ]
     ]
@@ -325,7 +325,7 @@ const getFirst = (first, filter, user, recursive) => {
 const getIndicesList = (filter) => {
   const limitType = getFilterValue(filter, 'type')
   const typeFilter = limitType
-    ? ({type}) => type === limitType
+    ? ({ type }) => type === limitType
     : Boolean
   const searchableFilter = ({ searchable = true }) => searchable
 
@@ -334,7 +334,10 @@ const getIndicesList = (filter) => {
 
 const hasFieldRequested = (fieldName, GraphQLResolveInfo) => {
   const fields = getFieldList(GraphQLResolveInfo, true)
-  return !!fields.find(field => field.indexOf(`.${fieldName}`) > -1)
+  return !!fields.find(field => (
+    field === fieldName ||
+    field.split('.').find(f => f === fieldName)
+  ))
 }
 
 const search = async (__, args, context, info) => {
@@ -354,12 +357,14 @@ const search = async (__, args, context, info) => {
   } = args
 
   // detect if Document.content is requested
-  let withoutContent
-  if (info) {
-    withoutContent = _withoutContent !== undefined ||
-      !hasFieldRequested('Document.content', info)
-  } else {
-    withoutContent = _withoutContent || false
+  let withoutContent = _withoutContent || false
+  let withoutChildren = withoutContent
+  if (info && _withoutContent === undefined) {
+    withoutContent = !hasFieldRequested('content', info)
+    withoutChildren = (
+      withoutContent &&
+      !hasFieldRequested('children', info)
+    )
   }
 
   const options = after
@@ -400,7 +405,7 @@ const search = async (__, args, context, info) => {
     index: indicesList.map(({ name }) => getIndexAlias(name, 'read')),
     from,
     size: first,
-    body: createQuery(search, filter, sort, indicesList, user, scheduledAt, withoutContent, withoutAggs, ignorePrepublished)
+    body: createQuery(search, filter, sort, indicesList, user, scheduledAt, withoutChildren, withoutAggs, ignorePrepublished)
   }
   debug('ES query', JSON.stringify(query))
 
@@ -444,7 +449,8 @@ const search = async (__, args, context, info) => {
   if (!recursive && !withoutRelatedDocs && (!filter.type || filter.type === 'Document')) {
     await addRelatedDocs({
       connection: response,
-      context
+      context,
+      withoutContent
     })
   }