only load linked docs in returned content or children

This speeds up getting the front document a lot. Locally the response time goes from ~140ms to ~70ms (both with hot caches) with the smaller dev documents collection. In prd the gains may reach 5x to 10x. I expect the current prd ~700ms to go down to ~200ms.

~ToDo~:
- ~new `afterDate` & `beforeDate` query for `Document.children` for overview page~

Decided not to implement advanced children filtering in backend (credit line based) and instead hard code some teaser ids in the frontend: https://github.com/orbiting/republik-frontend/commit/7bdab34cc56717c146003fdc5f75e16c3285dd58#diff-048df1fa60145e7b472e7ed8cafbcf72R37

This will allow for the same optimisations.